# perf: plugin load