--- a/ape_addressbook/__init__.py
+++ b/ape_addressbook/__init__.py
@@ -1,20 +1,31 @@
 from ape import plugins
-from ape.types import AddressType
-
-from .addressbook import AddressBookConfig, addressbook  # noqa: F401
-from .converters import AddressBookConverter
 
 
 @plugins.register(plugins.Config)
 def config_class():
+    from .addresses import AddressBookConfig
+
     return AddressBookConfig
 
 
 @plugins.register(plugins.ConversionPlugin)
 def converters():
+    from ape.types import AddressType
+
+    from .converters import AddressBookConverter
+
     return AddressType, AddressBookConverter
 
 
-___all__ = [
-    "addressbook",
-]
+def __getattr__(name: str):
+    if name == "addressbook":
+        from ape_addressbook.addresses import addressbook
+
+        return addressbook
+
+    import ape_addressbook.addresses as module
+
+    return getattr(module, name)
+
+
+___all__ = ["addressbook"]