--- a/ape_addressbook/converters.py
+++ b/ape_addressbook/converters.py
@@ -1,14 +1,16 @@
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 from ape.api.convert import ConverterAPI
-from ape.types import AddressType
 
-from .addressbook import addressbook
+from ape_addressbook import addressbook
+
+if TYPE_CHECKING:
+    from ape.types import AddressType
 
 
 class AddressBookConverter(ConverterAPI):
     def is_convertible(self, value: Any) -> bool:
         return isinstance(value, str) and value in addressbook
 
-    def convert(self, alias: str) -> AddressType:
+    def convert(self, alias: str) -> "AddressType":
         return addressbook[alias]