### What I did

Before:

```
In [1]: %time import ape_addressbook
CPU times: user 396 ms, sys: 50.4 ms, total: 446 ms
Wall time: 457 ms
```

After:

```
In [1]: %time import ape_addressbook
CPU times: user 6.89 ms, sys: 2.32 ms, total: 9.22 ms
Wall time: 8.14 ms
```

This one required changing the module name from `ape_addressbook.addressbook` to `ape_addressbook.addresses` because it made `from ape_addressbook import addressbook` always import the module instead of the instance.
There are other bugs when it comes to doing that I have seen in the past so even though it is a little breaky, it doesntt affect the public intended usage of this plugin and fixes a variety of issues so may be worth doing now under that context.

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)