@antazoey 2024-11-14
> age of this plugin and fixes a variety of issues so may be worth doing now under that context.

I am thinking of it as more of a bugfix so we are good, basically it was a bit of name collusion but Python didnt care before but gave you only 1  of the things and now it just works for either since the names are different, if that makes sense